+++ WebUI/Areas/Admin/Controllers/ServiceController.cs
@@ -1,4 +1,5 @@
 ﻿using DtoLayer.ServiceDtos;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using System.Text;
@@ -17,9 +18,15 @@ namespace WebUI.Areas.Admin.Controllers
             _httpClientFactory = httpClientFactory;
         }
 
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [Route("Index")]
         public async Task< IActionResult> Index()
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var responseMessage = await client.GetAsync("https://localhost:7250/api/Services");
             if (responseMessage.IsSuccessStatusCode)
@@ -31,16 +38,24 @@ namespace WebUI.Areas.Admin.Controllers
 
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpGet]
         [Route("CreateService")]
         public IActionResult CreateService()
         {
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         [Route("CreateService")]
         public async Task<IActionResult> CreateService(CreateServiceDto createServiceDto)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var jsonData = JsonConvert.SerializeObject(createServiceDto);
             var content = new StringContent(jsonData, Encoding.UTF8, "application/json");
@@ -51,10 +66,16 @@ namespace WebUI.Areas.Admin.Controllers
             }
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpDelete]
         [Route("RemoveService/{id}")]
         public async Task<IActionResult> RemoveService(int id)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var responseMessage = await client.DeleteAsync($"https://localhost:7250/api/Services/{id}");
             if (responseMessage.IsSuccessStatusCode)
@@ -63,10 +84,16 @@ namespace WebUI.Areas.Admin.Controllers
             }
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpGet]
         [Route("UpdateService/{id}")]
         public async Task<IActionResult> UpdateService(int id)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var responseMessage = await client.GetAsync($"https://localhost:7250/api/Services/{id}");
             if (responseMessage.IsSuccessStatusCode)
@@ -77,10 +104,16 @@ namespace WebUI.Areas.Admin.Controllers
             }
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         [Route("UpdateService/{id}")]
         public async Task<IActionResult> UpdateService(UpdateServiceDto updateServiceDto)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var jsonData = JsonConvert.SerializeObject(updateServiceDto);
             var content = new StringContent(jsonData, Encoding.UTF8, "application/json");
